feat: 5 новых профилей + антиИИ-пакет + исправление 2 багов в движке
Профили:
- it_pro: быстрая печать, оксфордская запятая, слэш вместо «или»
- romantic: многоточия, скобки-смайлы, пропущенные запятые
- newbie: пробелы перед знаками, промахи по ЙЦУКЕН
- gen_z: всё строчными, нет точек в конце, цифры вместо слов
- mobile_rush: слипание предложений, слияние абзацев

Базовый анти-ИИ пакет (5.2+5.5+5.7 = 100%) добавлен во все профили.

Исправления движка (найдены через 77 unit-тестов):
- 6.5: добавлен 'еще' в whitelist — после 5.7 (ё→е) слово переставало удваиваться
- 8.7: \b после группы 2 блокирует regex-backtrack в или...или конструкции

Сопутствующее (из предыдущих сессий):
- Поддержка дробных вероятностей (float) во всём стеке: _coin, registry, processor, widgets, cli

+++ engine/processor.py
@@ -45,7 +45,7 @@ APPLY_ORDER: list[str] = [
 
 
 class TextProcessor:
-    def __init__(self, profile: dict[str, int], seed: int | None = None,
+    def __init__(self, profile: dict[str, float], seed: int | None = None,
                  multiplier: float = 1.0):
         """
         profile: {marker_id: probability 0-100}
@@ -57,10 +57,10 @@ class TextProcessor:
         self.multiplier = multiplier
         self.stats: dict[str, int] = {}
 
-    def _effective_prob(self, marker_id: str) -> int:
+    def _effective_prob(self, marker_id: str) -> float:
         """Получить эффективную вероятность с учётом множителя."""
         base = self.profile.get(marker_id, 0)
-        return min(100, max(0, int(base * self.multiplier)))
+        return min(100.0, max(0.0, base * self.multiplier))
 
     def process(self, text: str) -> str:
         """Применяет все активные маркеры и возвращает модифицированный текст."""
